recipes-bs: u-boot: gen-container alignment is only for dolphin

+++ cookbook/recipes-bsp/u-boot/gen-container.sh
@@ -8,8 +8,10 @@ bootloader_append_size=`expr 512 - ${bootloader_subimg_size} % 512`
 
 cp ${_DEPLOY_DIR}/bootloader_nopreload.subimg ${_DEPLOY_DIR}/bootloader.subimg
 
-if [ ${bootloader_append_size} -lt 512 ]; then
-    dd if=/dev/zero of=${_DEPLOY_DIR}/bootloader.subimg bs=1 seek=${bootloader_subimg_size} count=${bootloader_append_size} conv=notrunc
+if [ "$MACHINE" == "winglet" ] || [ "$MACHINE" == "sl1680" ]; then
+    if [ ${bootloader_append_size} -lt 512 ]; then
+        dd if=/dev/zero of=${_DEPLOY_DIR}/bootloader.subimg bs=1 seek=${bootloader_subimg_size} count=${bootloader_append_size} conv=notrunc
+    fi
 fi
 
 if [ -f ${_DEPLOY_DIR}/preload_ta.subimg ];then
